--- astropy/io/fits/_tiled_compression/__init__.py
@@ -1,1 +1,0 @@
-from .tiled_compression import *
