+++ astropy/io/fits/hdu/compressed/_tiled_compression.py
@@ -10,8 +10,8 @@ import numpy as np
 
 from astropy.io.fits.hdu.base import BITPIX2DTYPE
 
-from .codecs import PLIO1, Gzip1, Gzip2, HCompress1, NoCompress, Rice1
-from .quantization import DITHER_METHODS, QuantizationFailedException, Quantize
+from ._codecs import PLIO1, Gzip1, Gzip2, HCompress1, NoCompress, Rice1
+from ._quantization import DITHER_METHODS, QuantizationFailedException, Quantize
 from .utils import _data_shape, _iter_array_tiles, _tile_shape
 
 ALGORITHMS = {
