+++ astropy/io/fits/hdu/compressed/section.py
@@ -2,8 +2,8 @@
 
 import numpy as np
 
-from astropy.io.fits._tiled_compression import decompress_image_data_section
-from astropy.io.fits._tiled_compression.utils import _data_shape, _n_tiles, _tile_shape
+from astropy.io.fits.hdu.compressed._tiled_compression import decompress_image_data_section
+from astropy.io.fits.hdu.compressed.utils import _data_shape, _n_tiles, _tile_shape
 from astropy.io.fits.hdu.base import BITPIX2DTYPE
 from astropy.utils.shapes import simplify_basic_index
