+++ astropy/io/fits/hdu/compressed/setup_package.py
@@ -10,7 +10,7 @@ SRC_DIR = os.path.join(os.path.dirname(__file__), "src")
 def get_extensions():
     return [
         Extension(
-            "astropy.io.fits._tiled_compression._compression",
+            "astropy.io.fits.hdu.compressed._compression",
             sources=[
                 os.path.join(SRC_DIR, "compression.c"),
                 os.path.join(SRC_DIR, "unquantize.c"),
