+++ astropy/io/fits/hdu/compressed/src/quantize.h
