+++ astropy/io/fits/hdu/compressed/src/ricecomp.h
