+++ astropy/io/fits/hdu/compressed/tests/data/compressed_with_nan.fits
