+++ astropy/io/fits/hdu/compressed/tests/data/m13_gzip.fits
