+++ astropy/io/fits/hdu/compressed/tests/test_fitsio.py
