+++ cextern/cfitsio/README.rst
@@ -1,3 +1,3 @@
 This directory only contains the small subset of files from CFITSIO which are
-required for the astropy.io.fits._tiled_compression package. All files are
+required for the astropy.io.fits.hdu.compressed._tiled_compression package. All files are
 copied verbatim from CFITSIO and can easily be updated if needed.
