+++ cextern/trim_cfitsio.sh
@@ -66,6 +66,6 @@ EOF
 
 cat <<EOF >cfitsio/README.rst
 This directory only contains the small subset of files from CFITSIO which are
-required for the astropy.io.fits._tiled_compression package. All files are
+required for the astropy.io.fits.hdu.compressed._tiled_compression package. All files are
 copied verbatim from CFITSIO and can easily be updated if needed.
 EOF
