+++ docs/nitpick-exceptions
@@ -17,7 +17,7 @@ py:class astropy.modeling.projections.Projection
 # astropy.io.fits
 py:class astropy.io.fits.hdu.base.ExtensionHDU
 py:class astropy.io.fits.util.NotifierMixin
-py:class astropy.io.fits._tiled_compression.codecs.Codec
+py:class astropy.io.fits.hdu.compressed._tiled_compression.codecs.Codec
 py:class numcodecs.abc.Codec
 
 # astropy.io.misc.yaml
